Refactor: changes after PR comments
